Move chat to its own package
