+++ app/src/main/java/com/OxGames/Pluvia/ui/screen/chat/ChatScreen.kt
@@ -1,4 +1,4 @@
-package com.OxGames.Pluvia.ui.screen.friends
+package com.OxGames.Pluvia.ui.screen.chat
 
 import android.content.res.Configuration
 import androidx.compose.foundation.layout.Arrangement
